+++ nextstep-di/src/main/java/nextstep/di/factory/ApplicationContext.java
@@ -2,7 +2,11 @@ package nextstep.di.factory;
 
 import nextstep.annotation.ComponentScan;
 import nextstep.annotation.Configuration;
+import nextstep.di.factory.definition.BeanDefinition;
 import nextstep.di.factory.exception.NotFoundComponentScanException;
+import nextstep.di.factory.scanner.BeanScanner;
+import nextstep.di.factory.scanner.ClasspathBeanScanner;
+import nextstep.di.factory.scanner.ConfigurationBeanScanner;
 import nextstep.stereotype.Controller;
 import nextstep.stereotype.Repository;
 import nextstep.stereotype.Service;
