+++ nextstep-di/src/main/java/nextstep/di/factory/definition/MethodDefinition.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.definition;
 
 import nextstep.di.factory.exception.InvalidBeanException;
 import org.slf4j.Logger;
