+++ nextstep-di/src/main/java/nextstep/di/factory/scanner/ClasspathBeanScanner.java
@@ -1,5 +1,7 @@
-package nextstep.di.factory;
+package nextstep.di.factory.scanner;
 
+import nextstep.di.factory.definition.BeanDefinition;
+import nextstep.di.factory.definition.ConstructorDefinition;
 import nextstep.di.factory.exception.DuplicateBeanException;
 import org.reflections.Reflections;
