+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/AnnotationHandlerMapping.java
@@ -3,8 +3,8 @@ package nextstep.mvc.tobe;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import nextstep.di.factory.ApplicationContext;
-import nextstep.di.factory.BeanDefinition;
 import nextstep.di.factory.BeanFactory;
+import nextstep.di.factory.definition.BeanDefinition;
 import nextstep.mvc.HandlerMapping;
 import nextstep.stereotype.Controller;
 import nextstep.web.annotation.RequestMapping;
